-Fixes from source code analyzizer, closes #1768

+++ scene/animation/tween.cpp
@@ -269,7 +269,7 @@ Variant Tween::_run_equation(InterpolateData& p_data) {
 	{
 
 	case Variant::BOOL:
-		result = ((int) _run_equation(p_data.trans_type, p_data.ease_type, p_data.elapsed - p_data.delay, (int) initial_val, (int) delta_val, p_data.times_in_sec)) >= 0.5;
+		result = ( _run_equation(p_data.trans_type, p_data.ease_type, p_data.elapsed - p_data.delay,  initial_val,  delta_val, p_data.times_in_sec)) >= 0.5;
 		break;
 
 	case Variant::INT:
